Fixed mknod errors(setup.sh::make install).
git-svn-id: https://svn.code.sf.net/p/iraf64/code/trunk/src@247 0682cb55-6b43-0410-8320-9b94b4336a91

+++ iraf/unix/scripts/setup.sh
@@ -526,16 +526,18 @@ EOF
   if [ ! $S = 0 ]; then
     exit $S
   fi
-  ( cd $DESTDIR/dev
-    echo "Creating /dev/imtli"
-    rm -f imtli ; mknod -m 777 imtli p
-    echo "Creating /dev/imtlo"
-    rm -f imt1o ; mknod -m 777 imt1o p
-    rm -f imt1 ; ln -s imt1o imt1
-  )
-  S=$?
-  if [ ! $S = 0 ]; then
-    exit $S
+  if [ "$DESTDIR" != "" -o "$USER" = "root" ]; then
+    ( cd $DESTDIR/dev
+      echo "Creating /dev/imtli"
+      rm -f imtli ; mknod -m 777 imtli p
+      echo "Creating /dev/imtlo"
+      rm -f imt1o ; mknod -m 777 imt1o p
+      rm -f imt1 ; ln -s imt1o imt1
+    )
+    S=$?
+    if [ ! $S = 0 ]; then
+      exit $S
+    fi
   fi
   #
   mkdir -p $DESTDIR/$BINDIR
